Remove duplicated tests files
Mandatory so coverage can be run on both directories
