+++ tests/functionnal/CommonITILValidation.php
@@ -273,4 +273,114 @@ class CommonITILValidation extends DbTestCase {
       $this->boolean($ticket->getFromDB($tid))->isTrue();
       $this->integer((int)$ticket->getField('global_validation'))->isEqualTo(\CommonITILValidation::WAITING);*/
    }
+
+   protected function testComputeValidationProvider(): array {
+      return [
+         // 100% validation required
+         [
+            'accepted'           => 0,
+            'refused'            => 0,
+            'validation_percent' => 100,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 10,
+            'refused'            => 0,
+            'validation_percent' => 100,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 90,
+            'refused'            => 0,
+            'validation_percent' => 100,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 100,
+            'refused'            => 0,
+            'validation_percent' => 100,
+            'result'             => \CommonITILValidation::ACCEPTED,
+         ],
+         [
+            'accepted'           => 0,
+            'refused'            => 10,
+            'validation_percent' => 100,
+            'result'             => \CommonITILValidation::REFUSED,
+         ],
+         // 50% validation required
+         [
+            'accepted'           => 0,
+            'refused'            => 0,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 10,
+            'refused'            => 0,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 50,
+            'refused'            => 0,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::ACCEPTED,
+         ],
+         [
+            'accepted'           => 0,
+            'refused'            => 10,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 0,
+            'refused'            => 50,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 0,
+            'refused'            => 60,
+            'validation_percent' => 50,
+            'result'             => \CommonITILValidation::REFUSED,
+         ],
+         // 0% validation required
+         [
+            'accepted'           => 0,
+            'refused'            => 0,
+            'validation_percent' => 0,
+            'result'             => \CommonITILValidation::WAITING,
+         ],
+         [
+            'accepted'           => 10,
+            'refused'            => 0,
+            'validation_percent' => 0,
+            'result'             => \CommonITILValidation::ACCEPTED,
+         ],
+         [
+            'accepted'           => 0,
+            'refused'            => 10,
+            'validation_percent' => 0,
+            'result'             => \CommonITILValidation::REFUSED,
+         ],
+      ];
+   }
+
+   /**
+    * @dataprovider testComputeValidationProvider
+    */
+   public function testComputeValidation(
+      int $accepted,
+      int $refused,
+      int $validation_percent,
+      int $result
+   ): void {
+      $test_result = \CommonITILValidation::computeValidation(
+         $accepted,
+         $refused,
+         $validation_percent
+      );
+
+      $this->integer($test_result)->isEqualTo($result);
+   }
 }
