+++ tests/functionnal/KnowbaseItem.php
@@ -281,4 +281,42 @@ class KnowbaseItem extends DbTestCase {
       ]);
       $this->integer($count)->isEqualTo(2);
    }
+
+   public function testGetForCategory() {
+      global $DB;
+
+      // Prepare mocks
+      $m_db = new \mock\DB();
+      $m_kbi = new \mock\KnowbaseItem();
+
+      // Mocked db request result
+      $it = new \ArrayIterator([
+         ['id' => '1'],
+         ['id' => '2'],
+         ['id' => '3'],
+      ]);
+      $this->calling($m_db)->request = $it;
+
+      // Ignore get fromDB
+      $this->calling($m_kbi)->getFromDB = true;
+
+      // True for call 1 & 3, false for call 2 and every following calls
+      $this->calling($m_kbi)->canViewItem[0] = false;
+      $this->calling($m_kbi)->canViewItem[1] = true;
+      $this->calling($m_kbi)->canViewItem[2] = false;
+      $this->calling($m_kbi)->canViewItem[3] = true;
+
+      // Replace global DB with mocked DB
+      $DB = $m_db;
+
+      // Expected : [1, 3]
+      $this->array(\KnowbaseItem::getForCategory(1, $m_kbi))
+         ->hasSize(2)
+         ->containsValues([1, 3]);
+
+      // Expected : [-1]
+      $this->array(\KnowbaseItem::getForCategory(1, $m_kbi))
+         ->hasSize(1)
+         ->contains(-1);
+   }
 }
